sync: adopt prichat security hardening and reproducible build base
Port work that lived only on the legacy private repo:

- main.rs: fail-closed startup. Refuse to boot when auth is not
  configured unless ALLOW_ANONYMOUS=true (outside development), and
  refuse to boot when FORCE_HTTP=true without FORCE_SECURE_COOKIES.
- config.rs / .env.example: ALLOW_ANONYMOUS opt-in (default false).
- handlers/ws_token.rs: fold three sequential DashMap reads into one
  atomic get() to close the room-state read race.
- handlers/auth.rs: stop logging session_id, CSRF token, and cookie
  headers via tracing::debug (PII / secret leakage into log sinks).
- jwt.rs / handlers/websocket.rs: widen ttl_secs from i64 to u64 to
  remove the negative-TTL footgun.
- Dockerfile: pin builder to rust:1.85-bookworm (was nightly), copy
  Cargo.lock, build with --locked for reproducibility.
- .gitignore: track root Cargo.lock going forward.
- Cargo.lock: commit the current resolved graph.

+++ src/handlers/websocket.rs
@@ -87,7 +87,7 @@ pub async fn ws_handler(
 
     // SECURITY: Single-use token enforcement (prevents replay attacks)
     // Token can only be used once within its validity window
-    if !state.consume_token(claims.jti, state.config.jwt_token_ttl_secs as u64) {
+    if !state.consume_token(claims.jti, state.config.jwt_token_ttl_secs) {
         tracing::warn!(
             "JWT token replay attempt detected: jti={}, room={}",
             claims.jti,
